# api/grpc/matreshka-be_api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "grpc/matreshka-be_api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "MatreshkaBeAPI"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/config/list": {
      "post": {
        "operationId": "MatreshkaBeAPI_ListConfigs",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/matreshka_be_apiListConfigsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/matreshka_be_apiListConfigsRequest"
            }
          }
        ],
        "tags": [
          "MatreshkaBeAPI"
        ]
      }
    },
    "/config/patch/{serviceName}": {
      "post": {
        "operationId": "MatreshkaBeAPI_PatchConfig",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/matreshka_be_apiPatchConfigResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "serviceName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/matreshka_be_apiMatreshkaBeAPIPatchConfigBody"
            }
          }
        ],
        "tags": [
          "MatreshkaBeAPI"
        ]
      }
    },
    "/config/{serviceName}": {
      "get": {
        "operationId": "MatreshkaBeAPI_GetConfig",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/matreshka_be_apiGetConfigResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "serviceName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "MatreshkaBeAPI"
        ]
      },
      "post": {
        "operationId": "MatreshkaBeAPI_PostConfig",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/matreshka_be_apiPostConfigResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "serviceName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/matreshka_be_apiMatreshkaBeAPIPostConfigBody"
            }
          }
        ],
        "tags": [
          "MatreshkaBeAPI"
        ]
      }
    },
    "/version": {
      "get": {
        "operationId": "MatreshkaBeAPI_ApiVersion",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/matreshka_be_apiApiVersionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "MatreshkaBeAPI"
        ]
      }
    }
  },
  "definitions": {
    "matreshka_be_apiApiVersionResponse": {
      "type": "object",
      "properties": {
        "version": {
          "type": "string"
        }
      }
    },
    "matreshka_be_apiAppInfo": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "version": {
          "type": "string"
        }
      }
    },
    "matreshka_be_apiGetConfigResponse": {
      "type": "object",
      "properties": {
        "config": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "matreshka_be_apiListConfigsRequest": {
      "type": "object",
      "properties": {
        "listRequest": {
          "$ref": "#/definitions/matreshka_be_apiListRequest"
        },
        "searchPattern": {
          "type": "string"
        }
      }
    },
    "matreshka_be_apiListConfigsResponse": {
      "type": "object",
      "properties": {
        "services": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/matreshka_be_apiAppInfo"
          }
        }
      }
    },
    "matreshka_be_apiListRequest": {
      "type": "object",
      "properties": {
        "limit": {
          "type": "integer",
          "format": "int64"
        },
        "offset": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "matreshka_be_apiMatreshkaBeAPIPatchConfigBody": {
      "type": "object",
      "properties": {
        "pathToValue": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
    "matreshka_be_apiMatreshkaBeAPIPostConfigBody": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "matreshka_be_apiPatchConfigResponse": {
      "type": "object"
    },
    "matreshka_be_apiPostConfigResponse": {
      "type": "object"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}